scripts/requirements-pptx.txt
python-pptx>=1.0.0
# Required only when using Aiquire_logo.svg (SVG → PNG for slides)
cairosvg>=2.7.0
